bot/bot-app/pyproject.toml
[tool.poetry]
name = "bot-app"
version = "0.1.0"
description = ""
authors = ["marianobrc <marianomartinez84@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
pycurl = "*"
boto3 = "*"
celery = { extras = ["sqs"], version = "~5.2.7" }
kombu = { extras = ["sqs"], version = "^5.2.4" }
slack-sdk = "^3.17.2"
django-model-utils = "^4.2.0"
pytest-django = "^4.5.2"
pytest-mock = "^3.8.2"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^4.0.1"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
